Refactor Ollama health and response utilities
Extracted Ollama health check and model validation logic into a dedicated utility class (OllamaHealthValidator) and moved response conversion logic to OllamaResponseConverter. Updated OllamaClient to delegate health/model checks and response formatting to these utilities, improving modularity and testability. Adjusted related tests to use the new structure.
